split SngFileParserPart #48
